requirements.txt
aniso8601==9.0.1
boto3==1.18.58
botocore==1.21.58
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
